refactor(web): split timer/report/behavior-tracking, lazy-load charts, drop dead content
- visual-timer: extract state hook, audio hook, and dial/preset/controls/
  companion subcomponents; run completion from the tick event (not the
  updater) so finish side-effects fire once and stay StrictMode-safe
- report: extract ReportControls/ReportDocument + state hook; load saved
  questions via lazy initializer; keep paid 90-day default
- behavior-tracking: extract week header / order bar
- weekly-chart + admin analytics: code-split recharts via React.lazy
- delete unused resources-data.fr + resources-components; drop unused
  react-day-picker dependency
