+++ apps/web/package.json
@@ -33,7 +33,6 @@
     "motion": "^12.38.0",
     "next-themes": "^0.4.6",
     "react": "^19.2.5",
-    "react-day-picker": "^9.14.0",
     "react-dom": "^19.2.5",
     "react-i18next": "^17.0.2",
     "react-qr-code": "^2.0.18",
