+++ apps/web/src/routes/_authenticated/admin-analytics/analytics-sections.tsx
@@ -1,13 +1,4 @@
-import {
-  ResponsiveContainer,
-  BarChart,
-  Bar,
-  XAxis,
-  YAxis,
-  CartesianGrid,
-  Tooltip,
-  Legend,
-} from "recharts";
+import { lazy, Suspense } from "react";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import type {
   DerivedKpis,
@@ -18,12 +9,13 @@ import type {
 } from "@/hooks/use-admin-analytics";
 import {
   EVENT_LABELS,
-  EVENT_COLORS,
   formatPercent,
   formatNumber,
   formatDuration,
   pivotByDay,
 } from "./analytics-format";
+
+const DailyChartImpl = lazy(() => import("./daily-chart-impl"));
 
 export function AlertsSection({ alerts }: { alerts: AnalyticsAlert[] }) {
   if (alerts.length === 0) return null;
@@ -394,24 +386,9 @@ export function DailyChartSection({
             Aucun événement sur la période.
           </div>
         ) : (
-          <ResponsiveContainer width="100%" height="100%">
-            <BarChart data={series}>
-              <CartesianGrid strokeDasharray="3 3" />
-              <XAxis dataKey="date" />
-              <YAxis allowDecimals={false} />
-              <Tooltip />
-              <Legend />
-              {Object.entries(EVENT_LABELS).map(([key, label]) => (
-                <Bar
-                  key={key}
-                  dataKey={key}
-                  name={label}
-                  fill={EVENT_COLORS[key]}
-                  stackId="a"
-                />
-              ))}
-            </BarChart>
-          </ResponsiveContainer>
+          <Suspense fallback={<div className="h-full" />}>
+            <DailyChartImpl series={series} />
+          </Suspense>
         )}
       </CardContent>
     </Card>
