+++ apps/web/src/routes/_authenticated/report/index.tsx
@@ -1,13 +1,7 @@
-import { createFileRoute, Link } from "@tanstack/react-router";
+import { createFileRoute } from "@tanstack/react-router";
 import { useEffect, useMemo, useState } from "react";
-import { useTranslation } from "react-i18next";
-import { Printer, ArrowLeft, Lock, MessageSquare, CalendarRange, Send, Star, ShieldAlert, Users } from "lucide-react";
 import { Button } from "@/components/ui/button";
-import { Badge } from "@/components/ui/badge";
 import { PageLoader } from "@/components/ui/page-loader";
-import { Label } from "@/components/ui/label";
-import { Textarea } from "@/components/ui/textarea";
-import { Input } from "@/components/ui/input";
 import { useChildren } from "@/hooks/use-children";
 import { useStats, type SymptomPoint, type StatsPeriod, type CustomDateRange } from "@/hooks/use-stats";
 import { useJournal } from "@/hooks/use-journal";
@@ -15,12 +9,12 @@ import { useBarkleySteps } from "@/hooks/use-barkley";
 import { useCrisisItems } from "@/hooks/use-crisis-list";
 import { useBillingStatus, useCheckout } from "@/hooks/use-billing";
 import { useUiStore } from "@/stores/ui-store";
-import { formatAgeRange, getChildEmoji } from "@/lib/utils";
-import { tagConfig } from "@/components/journal/journal-card-data";
-import type { JournalTag } from "@focusflow/validators";
-import { KpiBox } from "./kpi-box";
-import { Sparkline } from "./sparkline";
+import { getChildEmoji } from "@/lib/utils";
+import { ArrowLeft, Users } from "lucide-react";
 import { UpsellCard } from "./upsell-card";
+import { ReportControls } from "./report-controls";
+import { ReportDocument } from "./report-document";
+import type { SymptomSummary, BarkleyProgress } from "./report-document";
 
 export const Route = createFileRoute("/_authenticated/report/")({
   component: ReportPage,
@@ -40,14 +34,6 @@ const PERIOD_OPTIONS: Array<{ value: StatsPeriod; label: string; days: number }>
   { value: "quarter", label: "90 jours", days: 90 },
 ];
 
-function formatDate(d: Date | string) {
-  return new Date(d).toLocaleDateString("fr-FR", {
-    day: "numeric",
-    month: "long",
-    year: "numeric",
-  });
-}
-
 function average(values: number[]): number | null {
   if (values.length === 0) return null;
   return values.reduce((a, b) => a + b, 0) / values.length;
@@ -64,11 +50,49 @@ function trend(values: number[]): "up" | "down" | "stable" | null {
   return diff > 0 ? "up" : "down";
 }
 
-function trendLabel(t: "up" | "down" | "stable" | null): string {
-  if (t === "up") return "↗ en progression";
-  if (t === "down") return "↘ en baisse";
-  if (t === "stable") return "→ stable";
-  return "—";
+/** Consolidates the 9 useState calls in ReportContent into one place. */
+function useReportContentState(isActive: boolean, childId: string) {
+  const storageKey = `toko-report-questions-${childId}`;
+  // Free tier defaults to month; paid keeps quarter as the most clinically-useful window.
+  const [period, setPeriod] = useState<StatsPeriod>(
+    isActive ? "quarter" : "month"
+  );
+  const [customRange, setCustomRange] = useState<CustomDateRange>(() => {
+    const to = new Date().toISOString().split("T")[0]!;
+    const from = new Date(Date.now() - 90 * 24 * 60 * 60 * 1000).toISOString().split("T")[0]!;
+    return { from, to };
+  });
+  const [now] = useState(() => new Date());
+  // Load saved questions from localStorage on first mount via lazy initializer.
+  const [questions, setQuestions] = useState<string>(
+    () => localStorage.getItem(storageKey) ?? ""
+  );
+  const [emailTo, setEmailTo] = useState("");
+  const [emailSending, setEmailSending] = useState(false);
+  const [emailSent, setEmailSent] = useState(false);
+  const [pdfDownloading, setPdfDownloading] = useState(false);
+  const [showLockedHint, setShowLockedHint] = useState(false);
+
+  // Persist questions to localStorage whenever they change.
+  useEffect(() => {
+    if (questions) {
+      localStorage.setItem(storageKey, questions);
+    } else {
+      localStorage.removeItem(storageKey);
+    }
+  }, [questions, storageKey]);
+
+  return {
+    period, setPeriod,
+    customRange, setCustomRange,
+    now,
+    questions, setQuestions,
+    emailTo, setEmailTo,
+    emailSending, setEmailSending,
+    emailSent, setEmailSent,
+    pdfDownloading, setPdfDownloading,
+    showLockedHint, setShowLockedHint,
+  };
 }
 
 function ReportPage() {
@@ -154,7 +178,6 @@ function ReportPage() {
             onClick={() => window.print()}
             className="gap-2 shadow-sm"
           >
-            <Printer className="size-4" />
             Télécharger en PDF
           </Button>
         </div>
@@ -186,44 +209,19 @@ function ReportPage() {
 }
 
 function ReportContent({ childId, isActive }: { childId: string; isActive: boolean }) {
-  const { t } = useTranslation();
   const checkout = useCheckout();
-  // Free tier defaults to month (its widest free period); paid keeps quarter
-  // as the most-clinically-useful window.
-  const [period, setPeriod] = useState<StatsPeriod>(
-    isActive ? "quarter" : "month"
-  );
-  const [customRange, setCustomRange] = useState<CustomDateRange>(() => {
-    const to = new Date().toISOString().split("T")[0]!;
-    const from = new Date(Date.now() - 90 * 24 * 60 * 60 * 1000).toISOString().split("T")[0]!;
-    return { from, to };
-  });
-  const [now] = useState(() => new Date());
-  const [questions, setQuestions] = useState("");
-  const [emailTo, setEmailTo] = useState("");
-  const [emailSending, setEmailSending] = useState(false);
-  const [emailSent, setEmailSent] = useState(false);
-  const [pdfDownloading, setPdfDownloading] = useState(false);
-  const [showLockedHint, setShowLockedHint] = useState(false);
-
-  // Persist annotations per-child in localStorage
-  const storageKey = `toko-report-questions-${childId}`;
-  useEffect(() => {
-    // Persist changes to localStorage; on first mount (empty questions),
-    // load from storage instead of overwriting it.
-    if (!questions) {
-      const saved = localStorage.getItem(storageKey);
-      if (saved) {
-        setQuestions(saved);
-        return;
-      }
-    }
-    if (questions) {
-      localStorage.setItem(storageKey, questions);
-    } else {
-      localStorage.removeItem(storageKey);
-    }
-  }, [questions, storageKey]);
+  const state = useReportContentState(isActive, childId);
+  const {
+    period, setPeriod,
+    customRange, setCustomRange,
+    now,
+    questions, setQuestions,
+    emailTo, setEmailTo,
+    emailSending, setEmailSending,
+    emailSent, setEmailSent,
+    pdfDownloading, setPdfDownloading,
+    showLockedHint, setShowLockedHint,
+  } = state;
 
   const { data: children } = useChildren();
   const { data: stats, isLoading: statsLoading } = useStats(
@@ -240,11 +238,9 @@ function ReportContent({ childId, isActive }: { childId: string; isActive: boole
     ? { value: "custom" as const, label: "Personnalisé", days: Math.max(1, Math.ceil((new Date(customRange.to).getTime() - new Date(customRange.from).getTime()) / (24 * 60 * 60 * 1000))) }
     : PERIOD_OPTIONS.find((p) => p.value === period)!;
 
-  const symptomSummaries = useMemo(() => {
+  const symptomSummaries = useMemo((): SymptomSummary[] => {
     if (!stats?.symptoms) return [];
-    const keys = Object.keys(symptomLabels) as Array<
-      keyof Omit<SymptomPoint, "date">
-    >;
+    const keys = Object.keys(symptomLabels) as Array<keyof Omit<SymptomPoint, "date">>;
     return keys.map((key) => {
       const values = stats.symptoms
         .map((p) => p[key])
@@ -260,7 +256,6 @@ function ReportContent({ childId, isActive }: { childId: string; isActive: boole
     });
   }, [stats]);
 
-  // Journal entries filtered to the selected period
   const periodStart = useMemo(() => {
     if (period === "custom") return new Date(customRange.from);
     const d = new Date();
@@ -276,15 +271,13 @@ function ReportContent({ childId, isActive }: { childId: string; isActive: boole
   const journalHighlights = useMemo(() => {
     if (!journal) return [];
     const startMs = periodStart.getTime();
-    const endMs = periodEnd.getTime() + 24 * 60 * 60 * 1000; // include end day
+    const endMs = periodEnd.getTime() + 24 * 60 * 60 * 1000;
     return [...journal]
       .filter((e) => {
-        const t = new Date(e.date).getTime();
-        return t >= startMs && t <= endMs;
+        const ms = new Date(e.date).getTime();
+        return ms >= startMs && ms <= endMs;
       })
-      .sort(
-        (a, b) => new Date(b.date).getTime() - new Date(a.date).getTime()
-      )
+      .sort((a, b) => new Date(b.date).getTime() - new Date(a.date).getTime())
       .slice(0, 10);
   }, [journal, periodStart, periodEnd]);
 
@@ -297,8 +290,7 @@ function ReportContent({ childId, isActive }: { childId: string; isActive: boole
     [journalHighlights]
   );
 
-  // Barkley progress: completed steps sorted by number
-  const barkleyProgress = useMemo(() => {
+  const barkleyProgress = useMemo((): BarkleyProgress | null => {
     if (!barkleySteps || barkleySteps.length === 0) return null;
     const completed = [...barkleySteps]
       .filter((s) => s.completedAt)
@@ -307,9 +299,6 @@ function ReportContent({ childId, isActive }: { childId: string; isActive: boole
     return { completed, currentStep: maxCompleted + 1, total: 10 };
   }, [barkleySteps]);
 
-  // Server PDF download. Hits POST /api/report/pdf, streams the response as a
-  // blob, and triggers a browser download. Reserved to Famille (plan check
-  // happens server-side), free tier falls back to window.print().
   const handleDownloadPdf = async () => {
     if (!child) return;
     setPdfDownloading(true);
@@ -327,7 +316,6 @@ function ReportContent({ childId, isActive }: { childId: string; isActive: boole
         }),
       });
       if (!response.ok) {
-        // Fallback to native print if the server-side PDF is unavailable.
         window.print();
         return;
       }
@@ -347,7 +335,6 @@ function ReportContent({ childId, isActive }: { childId: string; isActive: boole
     }
   };
 
-  // Email send handler
   const handleSendEmail = async () => {
     if (!emailTo.trim() || !child) return;
     setEmailSending(true);
@@ -375,480 +362,41 @@ function ReportContent({ childId, isActive }: { childId: string; isActive: boole
 
   return (
     <div className="report-page mx-auto max-w-3xl">
-      {/* Print controls (hidden when printing) */}
-      <div className="report-controls mb-6 flex flex-col gap-4">
-        <div className="flex flex-col gap-3 sm:flex-row sm:items-start sm:justify-between">
-          <div>
-            <Link
-              to="/account"
-              className="inline-flex items-center gap-1.5 text-sm text-muted-foreground transition-colors hover:text-foreground"
-            >
-              <ArrowLeft className="size-3.5" />
-              Retour à mon compte
-            </Link>
-            <h1 className="mt-2 font-heading text-2xl font-semibold tracking-tight sm:text-3xl">
-              Carnet de consultation TDAH
-            </h1>
-            <p className="mt-1 text-sm text-muted-foreground">
-              {isActive
-                ? "Aperçu ci-dessous · imprimez ou enregistrez en PDF via votre navigateur."
-                : "À apporter à votre prochain rendez-vous · imprimez ou enregistrez en PDF via votre navigateur."}
-            </p>
-          </div>
-          <Button
-            size="lg"
-            onClick={isActive ? handleDownloadPdf : () => window.print()}
-            disabled={pdfDownloading}
-            className="gap-2 shadow-sm self-start sm:self-auto"
-          >
-            <Printer className="size-4" />
-            {pdfDownloading ? "Génération…" : "Télécharger en PDF"}
-          </Button>
-        </div>
-
-        {/* Period selector */}
-        <div className="flex flex-wrap items-center gap-2">
-          <span className="text-sm font-medium text-muted-foreground">
-            Période :
-          </span>
-          <div
-            role="tablist"
-            aria-label="Sélection de la période du rapport"
-            className="inline-flex rounded-lg border border-border/60 bg-background p-0.5 shadow-sm"
-          >
-            {PERIOD_OPTIONS.map((opt) => {
-              const locked = !isActive && opt.value === "quarter";
-              return (
-                <button
-                  key={opt.value}
-                  type="button"
-                  role="tab"
-                  aria-selected={period === opt.value}
-                  onClick={() => {
-                    if (locked) {
-                      setShowLockedHint(true);
-                      return;
-                    }
-                    setShowLockedHint(false);
-                    setPeriod(opt.value);
-                  }}
-                  className={
-                    "rounded-md px-3 py-1.5 text-sm font-medium transition-colors inline-flex items-center gap-1.5 " +
-                    (period === opt.value
-                      ? "bg-primary text-primary-foreground shadow-sm"
-                      : "text-muted-foreground hover:text-foreground")
-                  }
-                >
-                  {opt.label}
-                  {locked && <Lock className="size-3" aria-hidden />}
-                </button>
-              );
-            })}
-            <button
-              type="button"
-              role="tab"
-              aria-selected={period === "custom"}
-              onClick={() => {
-                if (!isActive) {
-                  setShowLockedHint(true);
-                  return;
-                }
-                setShowLockedHint(false);
-                setPeriod("custom");
-              }}
-              className={
-                "rounded-md px-3 py-1.5 text-sm font-medium transition-colors inline-flex items-center gap-1.5 " +
-                (period === "custom"
-                  ? "bg-primary text-primary-foreground shadow-sm"
-                  : "text-muted-foreground hover:text-foreground")
-              }
-            >
-              <CalendarRange className="size-3.5" />
-              Personnalisé
-              {!isActive && <Lock className="size-3" aria-hidden />}
-            </button>
-          </div>
-        </div>
-
-        {showLockedHint && !isActive && (
-          <div className="rounded-lg border border-primary/20 bg-primary/5 p-3 text-sm">
-            <p className="font-medium text-foreground">
-              Période 90 jours et personnalisée, réservées au plan Famille
-            </p>
-            <p className="mt-1 text-muted-foreground">
-              Suivez les tendances trimestrielles indispensables aux titrations
-              et aux dossiers MDPH.
-            </p>
-            <Button
-              size="sm"
-              className="mt-2"
-              onClick={() => checkout.mutate()}
-              disabled={checkout.isPending}
-            >
-              Essayer Famille 14 jours gratuits
-            </Button>
-          </div>
-        )}
-
-        {/* Custom date range picker */}
-        {period === "custom" && isActive && (
-          <div className="flex flex-col gap-3 sm:flex-row sm:flex-wrap sm:items-center">
-            <div className="flex items-center gap-2">
-              <Label htmlFor="report-from" className="text-sm whitespace-nowrap">Du</Label>
-              <Input
-                id="report-from"
-                type="date"
-                value={customRange.from}
-                max={customRange.to}
-                onChange={(e) => setCustomRange((r) => ({ ...r, from: e.target.value }))}
-                className="w-full sm:w-auto"
-              />
-            </div>
-            <div className="flex items-center gap-2">
-              <Label htmlFor="report-to" className="text-sm whitespace-nowrap">au</Label>
-              <Input
-                id="report-to"
-                type="date"
-                value={customRange.to}
-                min={customRange.from}
-                max={now.toISOString().split("T")[0]}
-                onChange={(e) => setCustomRange((r) => ({ ...r, to: e.target.value }))}
-                className="w-full sm:w-auto"
-              />
-            </div>
-          </div>
-        )}
-
-        {/* Annotations / questions au médecin */}
-        <div className="space-y-2">
-          <Label
-            htmlFor="report-questions"
-            className="flex items-center gap-1.5 text-sm"
-          >
-            <MessageSquare className="size-3.5" />
-            Vos questions au médecin
-            <span className="text-xs font-normal text-muted-foreground">
-              (apparaîtront en en-tête du PDF)
-            </span>
-          </Label>
-          <Textarea
-            id="report-questions"
-            value={questions}
-            onChange={(e) => setQuestions(e.target.value)}
-            placeholder="Ex. Les crises du soir se multiplient depuis 3 semaines. Faut-il ajuster le traitement ?"
-            rows={3}
-            className="resize-none"
-          />
-          <p className="text-xs text-muted-foreground/80">
-            Sauvegardé automatiquement dans votre navigateur, pour cet enfant.
-          </p>
-        </div>
-
-        {/* Email delivery, paid only */}
-        {isActive ? (
-          <div className="space-y-2">
-            <Label
-              htmlFor="report-email"
-              className="flex items-center gap-1.5 text-sm"
-            >
-              <Send className="size-3.5" />
-              Envoyer par email au médecin
-            </Label>
-            <div className="flex flex-col gap-2 sm:flex-row">
-              <Input
-                id="report-email"
-                type="email"
-                inputMode="email"
-                autoComplete="email"
-                value={emailTo}
-                onChange={(e) => setEmailTo(e.target.value)}
-                placeholder="docteur@cabinet-medical.fr"
-                className="w-full sm:max-w-xs"
-              />
-              <Button
-                variant="outline"
-                onClick={handleSendEmail}
-                disabled={emailSending || !emailTo.trim() || emailSent}
-                className="w-full gap-1.5 whitespace-nowrap sm:w-auto"
-              >
-                <Send className="size-3.5" />
-                {emailSent ? "Envoyé !" : emailSending ? "Envoi…" : "Envoyer"}
-              </Button>
-            </div>
-            <p className="text-xs text-muted-foreground/80">
-              Le rapport sera envoyé en pièce jointe HTML au médecin.
-            </p>
-          </div>
-        ) : null}
-      </div>
-
-      {/* Printable document */}
-      <article className="report-document rounded-xl border border-border/60 bg-card p-4 shadow-sm sm:p-6 md:p-10">
-        <header className="report-header border-b border-border/60 pb-6">
-          <div className="flex flex-col items-start gap-4 sm:flex-row sm:justify-between">
-            <div className="min-w-0">
-              <p className="text-xs font-medium uppercase tracking-wide text-primary">
-                Rapport TDAH · Tokō
-              </p>
-              <h2 className="mt-2 font-heading text-2xl font-semibold">
-                {child ? `${child.name}` : "Enfant"}
-              </h2>
-              <p className="mt-1 text-sm text-muted-foreground">
-                {child?.gender && (
-                  <span className="mr-2">
-                    {getChildEmoji(child.gender)}{" "}
-                    {child.gender === "male"
-                      ? "Garçon"
-                      : child.gender === "female"
-                        ? "Fille"
-                        : "—"}
-                  </span>
-                )}
-                {child?.ageRange && (
-                  <span>· {formatAgeRange(child.ageRange)}</span>
-                )}
-              </p>
-            </div>
-            <div className="text-left text-xs text-muted-foreground sm:text-right">
-              <p>
-                Période : {formatDate(periodStart)} → {formatDate(now)}
-              </p>
-              <p className="mt-1">Généré le {formatDate(now)}</p>
-            </div>
-          </div>
-        </header>
-
-        {/* Parent questions (if any) */}
-        {questions.trim() && (
-          <section className="report-section report-questions-box mt-6 rounded-lg border border-primary/20 bg-primary/5 p-4">
-            <h3 className="font-heading text-sm font-semibold uppercase tracking-wide text-primary">
-              Questions du parent
-            </h3>
-            <p className="mt-2 whitespace-pre-line text-sm leading-relaxed text-foreground/90">
-              {questions.trim()}
-            </p>
-          </section>
-        )}
-
-        {/* Overview */}
-        <section className="report-section mt-6">
-          <h3 className="font-heading text-base font-semibold uppercase tracking-wide text-muted-foreground">
-            Synthèse de la période ({periodConfig.label})
-          </h3>
-          <div className="mt-3 grid grid-cols-2 gap-3 sm:grid-cols-4">
-            <KpiBox
-              label="Entrées journal"
-              value={String(journalHighlights.length)}
-            />
-            <KpiBox
-              label="Jours suivis"
-              value={String(stats?.symptoms?.length ?? 0)}
-            />
-            <KpiBox label="Crises notées" value={String(crisisCount)} />
-            <KpiBox label="Victoires notées" value={String(victoryCount)} />
-          </div>
-        </section>
-
-        {/* Symptom averages with sparklines */}
-        <section className="report-section mt-8">
-          <h3 className="font-heading text-base font-semibold uppercase tracking-wide text-muted-foreground">
-            Moyennes par dimension (échelle 1-5)
-          </h3>
-          <div className="mt-3 overflow-x-auto rounded-lg border border-border/60">
-            <table className="w-full min-w-[34rem] text-sm">
-              <thead>
-                <tr className="border-b border-border/60 bg-muted/40">
-                  <th className="px-3 py-2 text-left font-medium text-muted-foreground">
-                    Dimension
-                  </th>
-                  <th className="px-3 py-2 text-center font-medium text-muted-foreground">
-                    Moyenne
-                  </th>
-                  <th className="px-3 py-2 text-center font-medium text-muted-foreground">
-                    Évolution
-                  </th>
-                  <th className="px-3 py-2 text-center font-medium text-muted-foreground">
-                    Tendance
-                  </th>
-                  <th className="px-3 py-2 text-right font-medium text-muted-foreground">
-                    Relevés
-                  </th>
-                </tr>
-              </thead>
-              <tbody>
-                {symptomSummaries.map((s, i) => (
-                  <tr key={s.key} className={i % 2 === 0 ? "" : "bg-muted/20"}>
-                    <td className="px-3 py-2 font-medium text-foreground">
-                      {s.label}
-                    </td>
-                    <td className="px-3 py-2 text-center">
-                      {s.average !== null ? s.average.toFixed(1) : "—"}
-                    </td>
-                    <td className="px-3 py-2 text-center text-primary">
-                      <Sparkline values={s.series} />
-                    </td>
-                    <td className="px-3 py-2 text-center text-muted-foreground">
-                      {trendLabel(s.trend)}
-                    </td>
-                    <td className="px-3 py-2 text-right text-muted-foreground">
-                      {s.samples}
-                    </td>
-                  </tr>
-                ))}
-              </tbody>
-            </table>
-          </div>
-          <p className="mt-2 text-xs text-muted-foreground">
-            Tendance calculée sur la 2<sup>de</sup> moitié de la période vs la
-            1<sup>re</sup>. Seuil de stabilité : ±0,3 point.
-          </p>
-        </section>
-
-        {/* Journal highlights */}
-        <section className="report-section mt-8">
-          <h3 className="font-heading text-base font-semibold uppercase tracking-wide text-muted-foreground">
-            Moments marquants du journal
-          </h3>
-          {journalHighlights.length === 0 ? (
-            <p className="mt-3 text-sm text-muted-foreground">
-              Aucune entrée de journal sur cette période.
-            </p>
-          ) : (
-            <ul className="mt-3 space-y-3">
-              {journalHighlights.map((entry) => (
-                <li
-                  key={entry.id}
-                  className="rounded-lg border border-border/60 bg-background/40 p-3"
-                >
-                  <div className="flex items-start justify-between gap-3">
-                    <div className="flex items-center gap-2">
-                      <span className="text-xs text-muted-foreground">
-                        {formatDate(entry.date)}
-                      </span>
-                    </div>
-                    {entry.tags && entry.tags.length > 0 && (
-                      <div className="flex flex-wrap gap-1 justify-end">
-                        {entry.tags.map((tag: string) => {
-                          const cfg = tagConfig[tag as JournalTag];
-                          return (
-                            <Badge
-                              key={tag}
-                              variant="outline"
-                              className="text-2xs"
-                            >
-                              {cfg ? t(cfg.labelKey) : tag}
-                            </Badge>
-                          );
-                        })}
-                      </div>
-                    )}
-                  </div>
-                  {entry.text && (
-                    <p className="mt-2 text-sm leading-relaxed text-foreground/90">
-                      {entry.text}
-                    </p>
-                  )}
-                </li>
-              ))}
-            </ul>
-          )}
-        </section>
-
-        {/* Barkley progress section */}
-        {barkleyProgress && (
-          <section className="report-section mt-8">
-            <h3 className="font-heading text-base font-semibold uppercase tracking-wide text-muted-foreground">
-              <span className="inline-flex items-center gap-1.5">
-                <Star className="size-4" />
-                Programme Barkley
-              </span>
-            </h3>
-            <div className="mt-3 rounded-lg border border-border/60 p-4">
-              <div className="flex items-center gap-3">
-                <div className="text-sm">
-                  <span className="font-semibold">{barkleyProgress.completed.length}</span>
-                  <span className="text-muted-foreground"> / {barkleyProgress.total} étapes complétées</span>
-                </div>
-                {barkleyProgress.currentStep <= barkleyProgress.total && (
-                  <Badge variant="outline" className="text-xs">
-                    Étape en cours : {barkleyProgress.currentStep}
-                  </Badge>
-                )}
-                {barkleyProgress.currentStep > barkleyProgress.total && (
-                  <Badge className="border-success-border bg-success-surface text-success-foreground text-xs">
-                    Programme terminé
-                  </Badge>
-                )}
-              </div>
-              {/* Progress bar */}
-              <div className="mt-3 h-2 w-full overflow-hidden rounded-full bg-muted">
-                <div
-                  className="h-full rounded-full bg-primary transition-all"
-                  style={{ width: `${(barkleyProgress.completed.length / barkleyProgress.total) * 100}%` }}
-                />
-              </div>
-              {/* Completed steps list */}
-              {barkleyProgress.completed.length > 0 && (
-                <ul className="mt-3 space-y-1">
-                  {barkleyProgress.completed.map((step) => (
-                    <li key={step.id} className="flex items-center gap-2 text-sm">
-                      <span className="text-success-foreground">✓</span>
-                      <span className="font-medium">Étape {step.stepNumber}</span>
-                      {step.completedAt && (
-                        <span className="text-xs text-muted-foreground">
-                          {formatDate(step.completedAt)}
-                        </span>
-                      )}
-                    </li>
-                  ))}
-                </ul>
-              )}
-            </div>
-          </section>
-        )}
-
-        {/* Crisis list preview */}
-        {crisisItems && crisisItems.length > 0 && (
-          <section className="report-section mt-8">
-            <h3 className="font-heading text-base font-semibold uppercase tracking-wide text-muted-foreground">
-              <span className="inline-flex items-center gap-1.5">
-                <ShieldAlert className="size-4" />
-                Liste de crise
-              </span>
-            </h3>
-            <p className="mt-1 text-xs text-muted-foreground">
-              Stratégies de régulation configurées par le parent ({crisisItems.length} élément{crisisItems.length > 1 ? "s" : ""})
-            </p>
-            <ul className="mt-3 space-y-2">
-              {crisisItems.map((item, i) => (
-                <li
-                  key={item.id}
-                  className="flex items-center gap-2.5 rounded-lg border border-border/60 bg-background/40 px-3 py-2 text-sm"
-                >
-                  <span className="flex size-5 shrink-0 items-center justify-center rounded-full bg-muted text-xs font-medium text-muted-foreground">
-                    {i + 1}
-                  </span>
-                  {item.emoji && <span>{item.emoji}</span>}
-                  <span className="text-foreground/90">{item.label}</span>
-                </li>
-              ))}
-            </ul>
-          </section>
-        )}
-
-        <footer className="report-footer mt-10 border-t border-border/60 pt-4 text-center text-xs text-muted-foreground">
-          <p>
-            Ce rapport est généré à partir des données saisies par le parent.
-            Il ne constitue pas un diagnostic médical.
-          </p>
-          <p className="mt-1">
-            {isActive
-              ? "Tokō · toko.app"
-              : "Généré gratuitement avec Tokō — toko.app"}
-          </p>
-        </footer>
-      </article>
-
+      <ReportControls
+        isActive={isActive}
+        period={period}
+        setPeriod={setPeriod}
+        customRange={customRange}
+        setCustomRange={setCustomRange}
+        now={now}
+        questions={questions}
+        setQuestions={setQuestions}
+        emailTo={emailTo}
+        setEmailTo={setEmailTo}
+        emailSending={emailSending}
+        emailSent={emailSent}
+        pdfDownloading={pdfDownloading}
+        showLockedHint={showLockedHint}
+        setShowLockedHint={setShowLockedHint}
+        onDownloadPdf={handleDownloadPdf}
+        onSendEmail={handleSendEmail}
+        checkout={checkout}
+      />
+      <ReportDocument
+        child={child}
+        isActive={isActive}
+        periodLabel={periodConfig.label}
+        periodStart={periodStart}
+        now={now}
+        questions={questions}
+        stats={stats}
+        symptomSummaries={symptomSummaries}
+        journalHighlights={journalHighlights}
+        crisisCount={crisisCount}
+        victoryCount={victoryCount}
+        barkleyProgress={barkleyProgress}
+        crisisItems={crisisItems}
+      />
       {!isActive && <UpsellCard />}
     </div>
   );
